requirements.txt
# Django Core
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3

# Authentication & Security
djangorestframework-simplejwt>=5.3
django-ratelimit>=4.1
cryptography>=41.0
python-dotenv>=1.0

# Async & Real-time
channels>=4.0
channels-redis>=4.2
daphne>=4.0
twisted[http2,tls]>=22.10.0
pyopenssl>=23.2.0
service_identity>=23.1.0

# Task Queue
celery>=5.3
django-celery-beat>=2.5
redis>=5.0

# LangChain & LangGraph (Workflow Execution)
langchain>=0.1
langgraph>=0.0.40
langchain-openai>=0.0.5
langchain-google-genai>=0.0.6
langchain-community>=0.0.20
mcp>=0.1.0

# Vector Store (for Inference Engine)
sentence-transformers>=2.2
faiss-cpu>=1.7
boto3>=1.34             # S3 index sync

# HTTP & API
httpx>=0.26
aiohttp>=3.9
requests>=2.31

# Database
psycopg2-binary>=2.9  # PostgreSQL

# Utilities
pydantic>=2.5
python-dateutil>=2.8

# Testing
pytest>=7.4
pytest-django>=4.7
pytest-asyncio>=0.23

# Code Execution Sandbox (for Code nodes)
RestrictedPython>=7.0
wasmtime>=12.0.0

adrf
Pillow>=10.0
bleach>=6.1
python-magic-bin>=0.4.14; sys_platform == "win32"
python-magic>=0.4.27; sys_platform != "win32"
pypdf>=3.17.0
PyPDF2>=3.0.1
python-pptx>=1.0.2

# Browser Automation
selenium>=4.0.0
webdriver-manager>=4.0.0

# SDKs
tavily-python>=0.3.0
firecrawl-py>=0.0.16
ddgs>=9.10.0
